Machine Learning Academic Project
Prostate cancer is a disease in which prostate cells
growing in the body are found as cancerous, it is known to
be the most common cancer in men. This project aims to
classify given dataset of prostate cancer to extract signicant
information involving biomarkers as they are useful in detecting
the existence and advancement of cancer in the body. Our dataset
comprises of two types of data namely clinical and gene data. We
initiate by collaborating these two datasets together and perform
feature selection using the univariate chi square and decision tree
methods. We have extracted 100/1000/2000 number of features
from a total of 60484 features. We have applied oversampling
using Smote Algorithm to avoid the imbalance class problem. To
obtain meaningful information from this given data, we have
performed different classication methods using 10 fold cross
validation like Nave Bayes, K Nearest Neighbour, Support Vector
Machine(SVM) with Radial Basis Function and Random Forest.
We then assess the individual performances of these different
classiers by metrices for evaluation, the results are compared to
obtain the best classier.